// backend/src/order/dto/order.dto.ts
export class TicketDto {
  film: string;
  session: string;
  daytime: string;
  row: number;
  seat: number;
  price: number;
}

export class CreateOrderDto {
  email: string;
  phone: string;
  tickets: TicketDto[];
}

export class OrderTicketDto extends TicketDto {
  id: string;
}

export class OrderResponseDto {
  total: number;
  items: OrderTicketDto[];
}